feat(contextSelector): component for ContextSelectorHeader
