+++ packages/context-selector/package.json
@@ -37,7 +37,9 @@
     "url": "https://github.com/equinor/fusion-react-components/issues"
   },
   "dependencies": {
+    "@equinor/fusion-react-icon": "^0.2.15",
     "@equinor/fusion-react-searchable-dropdown": "^0.2.0",
+    "@equinor/fusion-react-styles": "^0.5.5",
     "@equinor/fusion-react-utils": "^2.0.4"
   }
 }
